+++ packages/deployment-service/package.json
@@ -2,19 +2,19 @@
   "name": "deployment-service",
   "dependencies": {
     "@homeservenow/serverless-aws-handler": "^0.2.4",
+    "@reapit/utils": "workspace:packages/utils",
     "class-transformer": "^0.4.0",
     "class-validator": "^0.13.1",
     "jsonwebtoken": "^8.5.1"
   },
   "scripts": {
-    "start": "yarn publicKeyFetch && yarn build && yarn serverless offline start --runtime nodejs12.x --stage local",
-    "test": "echo '...skipping...'",
+    "start": "reapit-fetch-public-keys yarn build && yarn serverless offline start --runtime nodejs12.x --stage local",
+    "test": "reapit-fetch-public-keys --test && echo '...skipping...'",
     "build": "echo '...skipping...'",
     "lint": "concurrently \"prettier src/**/*.ts --write \"  \"eslint --cache --ext=ts,tsx src --fix\"",
     "release": "yarn publicKeyFetch && serverless deploy",
     "publish": "echo '...skipping...'",
-    "conf": "yarn config-manager --namespace cloud --entity deployment-service --name local --mode fetch",
-    "publicKeyFetch": "node src/fetch.public.keys.js"
+    "conf": "yarn config-manager --namespace cloud --entity deployment-service --name local --mode fetch"
   },
   "devDependencies": {
     "@reapit/config-manager": "workspace:packages/config-manager",
