+++ packages/deployment-service/src/functions/delete-deployment.ts
@@ -7,7 +7,7 @@ import {
 import * as service from '@/services/deployment'
 import { ownership } from '@/utils'
 import { connectSessionVerifyDecodeIdTokenWithPublicKeys, LoginIdentity } from '@reapit/connect-session'
-import publicKeys from './../../publicKeys.json'
+import publicKeys from './../../public-keys.json'
 /**
  * Delete a deployment
  */
