+++ packages/deployment-service/src/functions/get-deployment.ts
@@ -3,7 +3,7 @@ import { DeploymentModel } from '@/models'
 import * as service from '@/services/deployment'
 import { ownership } from '@/utils'
 import { connectSessionVerifyDecodeIdTokenWithPublicKeys, LoginIdentity } from '@reapit/connect-session'
-import publicKeys from './../../publicKeys.json'
+import publicKeys from './../../public-keys.json'
 
 /**
  * Get a deployment by id
