+++ packages/deployment-service/src/functions/index.ts
@@ -1,9 +1,9 @@
-export * from './create.pipeline'
-export * from './task.population'
-export * from './update.pipeline'
-export * from './get.pipeline'
-export * from './create.deployment'
-export * from './delete.deployment'
-export * from './update.deployment'
-export * from './get.deployment'
-export * from './paginate.deployment'
+export * from './create-pipeline'
+export * from './task-population'
+export * from './update-pipeline'
+export * from './get-pipeline'
+export * from './create-deployment'
+export * from './delete-deployment'
+export * from './update-deployment'
+export * from './get-deployment'
+export * from './paginate-deployment'
